Add failed CR message box
